+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaClusterTest.java
@@ -63,6 +63,7 @@ import io.strimzi.api.kafka.model.listener.NodePortListenerBootstrapOverride;
 import io.strimzi.api.kafka.model.listener.NodePortListenerBrokerOverride;
 import io.strimzi.api.kafka.model.listener.RouteListenerBrokerOverride;
 import io.strimzi.api.kafka.model.template.ContainerTemplate;
+import io.strimzi.api.kafka.model.template.ExternalTrafficPolicy;
 import io.strimzi.certs.OpenSslCertManager;
 import io.strimzi.kafka.oauth.server.ServerConfig;
 import io.strimzi.operator.cluster.KafkaVersionTestUtils;
@@ -87,6 +88,7 @@ import java.util.stream.Collectors;
 import static io.strimzi.test.TestUtils.LINE_SEPARATOR;
 import static io.strimzi.test.TestUtils.set;
 import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
@@ -725,6 +727,8 @@ public class KafkaClusterTest {
         assertThat(ext.getSpec().getSelector(), is(kc.getSelectorLabels()));
         assertThat(ext.getSpec().getPorts(), is(Collections.singletonList(kc.createServicePort(KafkaCluster.EXTERNAL_PORT_NAME, KafkaCluster.EXTERNAL_PORT, KafkaCluster.EXTERNAL_PORT, "TCP"))));
         assertThat(ext.getSpec().getLoadBalancerIP(), is(nullValue()));
+        assertThat(ext.getSpec().getExternalTrafficPolicy(), is(nullValue()));
+        assertThat(ext.getSpec().getLoadBalancerSourceRanges(), is(emptyList()));
         checkOwnerReference(kc.createOwnerReference(), ext);
 
         // Check per pod services
@@ -735,7 +739,153 @@ public class KafkaClusterTest {
             assertThat(srv.getSpec().getSelector().get(Labels.KUBERNETES_STATEFULSET_POD_LABEL), is(KafkaCluster.kafkaPodName(cluster, i)));
             assertThat(srv.getSpec().getPorts(), is(Collections.singletonList(kc.createServicePort(KafkaCluster.EXTERNAL_PORT_NAME, KafkaCluster.EXTERNAL_PORT, KafkaCluster.EXTERNAL_PORT, "TCP"))));
             assertThat(srv.getSpec().getLoadBalancerIP(), is(nullValue()));
+            assertThat(srv.getSpec().getExternalTrafficPolicy(), is(nullValue()));
+            assertThat(srv.getSpec().getLoadBalancerSourceRanges(), is(emptyList()));
             checkOwnerReference(kc.createOwnerReference(), srv);
+        }
+    }
+
+    @Test
+    public void testLoadBalancerExternalTrafficPolicyLocal() {
+        Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
+                image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
+                .editSpec()
+                    .editKafka()
+                        .withNewListeners()
+                            .withNewKafkaListenerExternalLoadBalancer()
+                            .endKafkaListenerExternalLoadBalancer()
+                        .endListeners()
+                        .withNewTemplate()
+                            .withNewExternalBootstrapService()
+                                .withExternalTrafficPolicy(ExternalTrafficPolicy.LOCAL)
+                            .endExternalBootstrapService()
+                            .withNewPerPodService()
+                                .withExternalTrafficPolicy(ExternalTrafficPolicy.LOCAL)
+                            .endPerPodService()
+                        .endTemplate()
+                    .endKafka()
+                .endSpec()
+                .build();
+        KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);
+
+        // Check external bootstrap service
+        Service ext = kc.generateExternalBootstrapService();
+        assertThat(ext.getSpec().getExternalTrafficPolicy(), is(ExternalTrafficPolicy.LOCAL.toValue()));
+
+        // Check per pod services
+        for (int i = 0; i < replicas; i++)  {
+            Service srv = kc.generateExternalService(i);
+            assertThat(srv.getSpec().getExternalTrafficPolicy(), is(ExternalTrafficPolicy.LOCAL.toValue()));
+        }
+    }
+
+    @Test
+    public void testLoadBalancerExternalTrafficPolicyCluster() {
+        Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
+                image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
+                .editSpec()
+                    .editKafka()
+                        .withNewListeners()
+                            .withNewKafkaListenerExternalLoadBalancer()
+                            .endKafkaListenerExternalLoadBalancer()
+                        .endListeners()
+                        .withNewTemplate()
+                            .withNewExternalBootstrapService()
+                                .withExternalTrafficPolicy(ExternalTrafficPolicy.CLUSTER)
+                            .endExternalBootstrapService()
+                            .withNewPerPodService()
+                                .withExternalTrafficPolicy(ExternalTrafficPolicy.CLUSTER)
+                            .endPerPodService()
+                        .endTemplate()
+                    .endKafka()
+                .endSpec()
+                .build();
+        KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);
+
+        // Check external bootstrap service
+        Service ext = kc.generateExternalBootstrapService();
+        assertThat(ext.getSpec().getExternalTrafficPolicy(), is(ExternalTrafficPolicy.CLUSTER.toValue()));
+
+        // Check per pod services
+        for (int i = 0; i < replicas; i++)  {
+            Service srv = kc.generateExternalService(i);
+            assertThat(srv.getSpec().getExternalTrafficPolicy(), is(ExternalTrafficPolicy.CLUSTER.toValue()));
+        }
+    }
+
+    @Test
+    public void testLoadBalancerSourceRange() {
+        List<String> sourceRanges = new ArrayList();
+        sourceRanges.add("10.0.0.0/8");
+        sourceRanges.add("130.211.204.1/32");
+
+        Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
+                image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
+                .editSpec()
+                    .editKafka()
+                        .withNewListeners()
+                            .withNewKafkaListenerExternalLoadBalancer()
+                            .endKafkaListenerExternalLoadBalancer()
+                        .endListeners()
+                        .withNewTemplate()
+                            .withNewExternalBootstrapService()
+                                .withLoadBalancerSourceRanges(sourceRanges)
+                            .endExternalBootstrapService()
+                            .withNewPerPodService()
+                                .withLoadBalancerSourceRanges(sourceRanges)
+                            .endPerPodService()
+                        .endTemplate()
+                    .endKafka()
+                .endSpec()
+                .build();
+        KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);
+
+        // Check external bootstrap service
+        Service ext = kc.generateExternalBootstrapService();
+        assertThat(ext.getSpec().getLoadBalancerSourceRanges(), is(sourceRanges));
+
+        // Check per pod services
+        for (int i = 0; i < replicas; i++)  {
+            Service srv = kc.generateExternalService(i);
+            assertThat(srv.getSpec().getLoadBalancerSourceRanges(), is(sourceRanges));
+        }
+    }
+
+    @Test
+    public void testSourceRangeOnNodePort() {
+        List<String> sourceRanges = new ArrayList();
+        sourceRanges.add("10.0.0.0/8");
+        sourceRanges.add("130.211.204.1/32");
+
+        Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
+                image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
+                .editSpec()
+                    .editKafka()
+                        .withNewListeners()
+                            .withNewKafkaListenerExternalNodePort()
+                            .endKafkaListenerExternalNodePort()
+                        .endListeners()
+                        .withNewTemplate()
+                            .withNewExternalBootstrapService()
+                                .withLoadBalancerSourceRanges(sourceRanges)
+                            .endExternalBootstrapService()
+                            .withNewPerPodService()
+                                .withLoadBalancerSourceRanges(sourceRanges)
+                            .endPerPodService()
+                        .endTemplate()
+                    .endKafka()
+                .endSpec()
+                .build();
+        KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);
+
+        // Check external bootstrap service
+        Service ext = kc.generateExternalBootstrapService();
+        assertThat(ext.getSpec().getLoadBalancerSourceRanges(), is(emptyList()));
+
+        // Check per pod services
+        for (int i = 0; i < replicas; i++)  {
+            Service srv = kc.generateExternalService(i);
+            assertThat(srv.getSpec().getLoadBalancerSourceRanges(), is(emptyList()));
         }
     }
 
@@ -744,13 +894,13 @@ public class KafkaClusterTest {
         Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
                 image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
                 .editSpec()
-                    .editKafka()
-                        .withNewListeners()
-                            .withNewKafkaListenerExternalLoadBalancer()
-                                .withTls(false)
-                            .endKafkaListenerExternalLoadBalancer()
-                        .endListeners()
-                    .endKafka()
+                .editKafka()
+                .withNewListeners()
+                .withNewKafkaListenerExternalLoadBalancer()
+                .withTls(false)
+                .endKafkaListenerExternalLoadBalancer()
+                .endListeners()
+                .endKafka()
                 .endSpec()
                 .build();
         KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);
@@ -789,16 +939,16 @@ public class KafkaClusterTest {
         Kafka kafkaAssembly = new KafkaBuilder(ResourceUtils.createKafkaCluster(namespace, cluster, replicas,
                 image, healthDelay, healthTimeout, metricsCm, configuration, emptyMap()))
                 .editSpec()
-                    .editKafka()
-                        .withNewListeners()
-                            .withNewKafkaListenerExternalLoadBalancer()
-                                .withNewOverrides()
-                                    .withBootstrap(bootstrapOverride)
-                                    .withBrokers(brokerOverride0, brokerOverride2)
-                                .endOverrides()
-                            .endKafkaListenerExternalLoadBalancer()
-                        .endListeners()
-                    .endKafka()
+                .editKafka()
+                .withNewListeners()
+                .withNewKafkaListenerExternalLoadBalancer()
+                .withNewOverrides()
+                .withBootstrap(bootstrapOverride)
+                .withBrokers(brokerOverride0, brokerOverride2)
+                .endOverrides()
+                .endKafkaListenerExternalLoadBalancer()
+                .endListeners()
+                .endKafka()
                 .endSpec()
                 .build();
         KafkaCluster kc = KafkaCluster.fromCrd(kafkaAssembly, VERSIONS);