+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add support for tini
 * When not explicitly configured by the user in `jvmOptions`, `-Xmx` option is calculated from memory requests rather than from memory limits
 * Expose JMX port on Kafka brokers via an internal service
+* Add support for `externalTrafficPolicy` and `loadBalancerSourceRanges` properties on loadbalancer and nodeport type services
 * Add support for user quotas
 * Add support for Istio protocol selection in service port names  
 Note: Strimzi is essentially adding a `tcp-` prefix to the port names in Kafka services and headless services.